pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "synccore"
version = "1.1.0"
description = "One-way client/server file synchronisation over HTTPS"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "watchdog>=4.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "typer[all]>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "rich>=13.7",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
]

[project.scripts]
synccore = "main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
